Refactor complete_bipartite_graph. Fixes #1375.
This commit reimplements `complete_bipartite_graph` using a generator
expression instead of two nested for loops. Also adds the `bipartite`
node attribute (that was missing, as per #1375).

I've also moved the `complete_bipartite_graph` function from `classic.py`
to `bipartite.py`. I think it's better to have all bipartite generators
together.

The function `complete_bipartite_graph` is used by other functions, such as
`wheel_graph`, which was abusing the interface by passing -1 as the second
argument of `complete_bipartite_graph` (for `wheel_graph(0)`) and hoping to
get an empty graph with zero nodes. This is also fixed.
